Fix relay's tests
